fix profile model
