+++ main_app/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import Post, Profile
+from .models import Post
 
 # Register your models here.
 admin.site.register(Post)
-admin.site.register(Profile)
+# admin.site.register(Profile)
