+++ userprofile/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Profile
+# Register your models here.
+admin.site.register(Profile)
